+++ src/sentry/conf/server.py
@@ -723,6 +723,7 @@ CELERY_IMPORTS = (
     "sentry.tasks.weekly_escalating_forecast",
     "sentry.tasks.auto_ongoing_issues",
     "sentry.tasks.auto_archive_issues",
+    "sentry.tasks.check_am2_compatibility",
 )
 
 CELERY_QUEUES = [
